requirements.txt
fastapi
uvicorn
gunicorn
pymongo
python-jose
passlib[argon2]
python-dotenv